fix: 영수증 프린트 함수 수정
